requirements.txt
django==5.1.6
djangorestframework==3.15.2
django-phonenumbers==1.0.1
django-phonenumber-field==8.0.0
django_extensions==3.2.3

# documentaion
mkdocs
mkdocs-material
mkdocs-git-revision-date-localized-plugin
